Sixteenth Attempt of HW1

+++ FVCalc.java
@@ -5,7 +5,7 @@ public class FVCalc {
         // Parse arguments and convert to appropriate types
         int p = Integer.parseInt(args[0]); // Initial investment amount (Principal)
         
-        int r = Integer.parseInt(args[1]); // Interest rate as an integer (e.g., 5 for 5%)
+        double r = Double.parseDouble(args[1]); // Interest rate as an integer (e.g., 5 for 5%)
         r = r / 100; // Convert from percentage to decimal
 
         int y = Integer.parseInt(args[2]); // Number of years for the investment
